Created and tested the structure macros.
This code is a test at building strucre macros to simplify the
branching with my 65816 code. This allows for the following
strctures:
	* IF_CONDITION ... [ ELSE ... ] ENDIF
	* REPEAT ... UNTIL_CONDITION
	* REPEAT ... FOREVER
	* FOR_XY start, direction, end ... NEXT
	* CONTINUE for all loops
	* BREAK for all loops

The structures are maintained using an 8 layer numeric variable stack,
allowing for the structures to be nested.

It has been testing using the `examples/flashing_colors.s` code.

The following files were sourced from a previous project:
  * includes/registers.inc
  * includes/sfc_header.inc
  * config/LOROM_1MBit.cfg