Import/export use __ instead of ::
This fixes a number of issues with getting the `Text_PrintDecimal`
macro to work properly. Previously the assembler was preventing
me from using .ident with a `::`. Now that there are no `::` the
problem is fixed.

Also, according to the ca65 mailing list, there is a bug involving
zp size/type when using scopes. As there are now no scopes, the
chance of a zp variable being mistaken for an addr are now limited.