Updated Metatile1x16 so it uses Structure of Arrays for tiles.
This means that the value of the map is (tile * 2) NOT (tile * 8),
this will make the tile-physics portion easier because the tile
physics data doesn't fit in 8 bytes anyway.

This would also save space as most tiles are identical (according
to physics) and it would be a waste of space to use 8 bytes to store
thir data instead of a 2 byte pointer.